Implement a hex protocol for data transfer
This implements a hex message protocol with the following structure:

2 hex chars: hat state
4 hex chars: button state
2 hex chars: lx
2 hex chars: ly
2 hex chars: rx
2 hex chars: ry
terminated with \n or \n (or both).

For a total of 15 bytes per message.

Upper or lower case hex characters are accepted. Non-hex characters
will be completely ignored.

Early or late termination will cause the whole message up to this
point to be dropped.

The button state is big endian, ie the high byte comes first. The rest
is identical to the HID report format, except hex encoded.

Parsing code has been moved out of the ISR and replaced by a circular
buffer to avoid overruns.

The code will never send back a response on serial unless there is
an overrun, then it will send "X".

This code is stable at 115200 bps.