Add python SDL to Switch bridge
This code is incomplete. It uses the SDL Game Controller subsystem
which is easy to use because it makes all controllers look like
an XBOX 360 controller. Unfortunately it doesn't have enough buttons
for every switch button.

See "./bridge.py --help" for usage. You can specify the controller
by number or name, and set the port and baudrate. The default baud
rate matches the one in Joystick.c as of this commit. The other
defaults should work if you have only one controller.

Rather than sending an update for every event, this sends an update
every ~5 msec (subject to python timing inaccuracy). It seems very
responsive compared to previous versions.

In order to get the full joystick mapping working the code should
be modified to use SDL_Joystick instead of SDL_GameController
however this code will be joystick-specific as there is no standard
mapping for SDL_Joystick.