Synchronise the bridge with the Switch interrupt transfers
Each time the Arduino revieves a transfer request on the IN end
point, write 'U' to the serial port. The bridge waits for 'U' after
sending an update.

This synchronizes the flow of information from the PC to the switch
such that it should only send one controller update each time the
Switch polls for information on USB.

This is not guaranteed to be related to frames at all because the
USB host controller chip will be running independent of the main
CPU.